Delete Coursera Answers
Why are you applying for Financial Aid? (150 words minimum required)

I am a student and not an earner. I am a CSE Undergrad. My family annual income ranges from 166-167$(in normal conditions) and due to this pandemic it has reduced to around too less. My father is the only earning member in the house. Its impossible to make both ends meet with this minimal income. Thankfully , I got a scholarship which helped me get an admission to college. I go for 10+ hours coaching in my campus so that I can help my parents in this condition. Its too tough in this condition. So this financial is really needed so that I can get a good job at a product based or service based company and improve my present and future.
If I add up my course it will make even harder for them to pay. Sir, I need this course very badly for my CV and for increasing my knowledge about the subject.

How will taking this course help you achieve your career goals? (150 words minimum required)

Since I am a CSE Undergrad , projects are needed to build a portfloi. And to build  a worthy project is a very big task. I want to complete the course due to my curiosity and also that I can put a good CV to a get applied to a job. This Course will boost my job prospects after graduation from my institute. It will help perform better in carrying out various projects in a different languages and give me an edge over my competitors. A verified certificate will attach credibility to the certificate I receive from this course. I plan to complete all assignments on or before time as I have done in previous Signature Track Courses. Also I intend to participate in Discussion Forums, which I have found to supplement my learning immensely in the other online courses I have taken on Coursera. I also plan to grade assignments which are to peer reviewed which I believe will an invaluable learning opportunity.


Would you consider using a low-interest loan to pay for your courses?If you answered no, please help us understand why.

Sir, the financial status of the family is not too good to pay the loan amount. We already have a lot of dept in the bank and my parents are paying it on regular basis. It would make their life even harder to add a new loan over them. Sir, i don't want to put any pressure over them. Sir, it would be a great help for me to get a good job and help my family if I'm able to get this course.